Simply the use of custom codepoint regions
If an application is already using parts of the Private Use Areas, it
can be cumbersome to use the full constructor just to pass in a
baseCodePoint to the EmojiManager. This way, we can pass in either a
standard region (via the new StandardUnassignedUnicodeRegion enum) or a
completely custom integer.